+++ tfe/data_source_workspace_ids_test.go
@@ -73,7 +73,7 @@ func TestAccTFEWorkspaceIDsDataSource_wildcard(t *testing.T) {
 		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt),
+				Config: testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt, "*"),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					// names attribute
 					resource.TestCheckResourceAttr(
@@ -113,6 +113,144 @@ func TestAccTFEWorkspaceIDsDataSource_wildcard(t *testing.T) {
 						"data.tfe_workspace_ids.foobar", fmt.Sprintf("ids.workspace-bar-%d", rInt)),
 					resource.TestCheckResourceAttrSet(
 						"data.tfe_workspace_ids.foobar", fmt.Sprintf("ids.workspace-dummy-%d", rInt)),
+
+					// id attribute
+					resource.TestCheckResourceAttrSet("data.tfe_workspace_ids.foobar", "id"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFEWorkspaceIDsDataSource_prefixWildcard(t *testing.T) {
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+	fooWorkspaceName := fmt.Sprintf("*-foo-%d", rInt)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt, fooWorkspaceName),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					// names attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "names.#", "1"),
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "names.0", fooWorkspaceName),
+
+					// organization attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "organization", orgName),
+
+					// full_names attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "full_names.%", "1"),
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar",
+						fmt.Sprintf("full_names.workspace-foo-%d", rInt),
+						fmt.Sprintf("tst-terraform-%d/workspace-foo-%d", rInt, rInt),
+					),
+
+					// ids attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "ids.%", "1"),
+					resource.TestCheckResourceAttrSet(
+						"data.tfe_workspace_ids.foobar", fmt.Sprintf("ids.workspace-foo-%d", rInt)),
+
+					// id attribute
+					resource.TestCheckResourceAttrSet("data.tfe_workspace_ids.foobar", "id"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFEWorkspaceIDsDataSource_suffixWildcard(t *testing.T) {
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+	fooWorkspaceName := "workspace-foo-*"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt, fooWorkspaceName),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					// names attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "names.#", "1"),
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "names.0", fooWorkspaceName),
+
+					// organization attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "organization", orgName),
+
+					// full_names attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "full_names.%", "1"),
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar",
+						fmt.Sprintf("full_names.workspace-foo-%d", rInt),
+						fmt.Sprintf("tst-terraform-%d/workspace-foo-%d", rInt, rInt),
+					),
+
+					// ids attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "ids.%", "1"),
+					resource.TestCheckResourceAttrSet(
+						"data.tfe_workspace_ids.foobar", fmt.Sprintf("ids.workspace-foo-%d", rInt)),
+
+					// id attribute
+					resource.TestCheckResourceAttrSet("data.tfe_workspace_ids.foobar", "id"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFEWorkspaceIDsDataSource_substringWildcard(t *testing.T) {
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+	fooWorkspaceName := "*-foo-*"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt, fooWorkspaceName),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					// names attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "names.#", "1"),
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "names.0", fooWorkspaceName),
+
+					// organization attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "organization", orgName),
+
+					// full_names attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "full_names.%", "1"),
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar",
+						fmt.Sprintf("full_names.workspace-foo-%d", rInt),
+						fmt.Sprintf("tst-terraform-%d/workspace-foo-%d", rInt, rInt),
+					),
+
+					// ids attribute
+					resource.TestCheckResourceAttr(
+						"data.tfe_workspace_ids.foobar", "ids.%", "1"),
+					resource.TestCheckResourceAttrSet(
+						"data.tfe_workspace_ids.foobar", fmt.Sprintf("ids.workspace-foo-%d", rInt)),
 
 					// id attribute
 					resource.TestCheckResourceAttrSet("data.tfe_workspace_ids.foobar", "id"),
@@ -357,7 +495,7 @@ data "tfe_workspace_ids" "foobar" {
 }`, rInt, rInt, rInt, rInt)
 }
 
-func testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt int) string {
+func testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt int, wildcardName string) string {
 	return fmt.Sprintf(`
 resource "tfe_organization" "foobar" {
   name  = "tst-terraform-%d"
@@ -380,14 +518,14 @@ resource "tfe_workspace" "dummy" {
 }
 
 data "tfe_workspace_ids" "foobar" {
-  names        = ["*"]
+  names        = ["%s"]
   organization = tfe_workspace.dummy.organization
   depends_on = [
     tfe_workspace.foo,
     tfe_workspace.bar,
     tfe_workspace.dummy
   ]
-}`, rInt, rInt, rInt, rInt)
+}`, rInt, rInt, rInt, rInt, wildcardName)
 }
 
 func testAccTFEWorkspaceIDsDataSourceConfig_tags(rInt int) string {